[REFACT] ISO 포맷 변환 메서드 리팩토링
