+++ src/main/java/com/sopterm/makeawish/common/Util.java
@@ -1,5 +1,12 @@
 package com.sopterm.makeawish.common;
 
+import java.time.Instant;
+import java.time.LocalDateTime;
+import java.time.ZoneId;
+import java.time.format.DateTimeFormatter;
+
+import lombok.val;
+
 public class Util {
 
 	private static final float FEE = 3.4f;
@@ -10,5 +17,12 @@ public class Util {
 
 	public static int getPricePercent(int totalPrice, int presentPrice) {
 		return (getPriceAppliedFee(totalPrice) / presentPrice) * 100;
+	}
+
+	public static LocalDateTime convertToTime(String date) {
+		val instant = Instant
+			.from(DateTimeFormatter.ISO_DATE_TIME.parse(date))
+			.atZone(ZoneId.of("Asia/Seoul"));
+		return LocalDateTime.from(instant);
 	}
 }
