+++ src/main/java/com/sopterm/makeawish/dto/wish/WishRequestDTO.java
@@ -1,11 +1,6 @@
 package com.sopterm.makeawish.dto.wish;
 
-import java.time.Instant;
-import java.time.LocalDateTime;
-import java.time.ZoneId;
-import java.time.format.DateTimeFormatter;
-
-import lombok.val;
+import static com.sopterm.makeawish.common.Util.*;
 
 import com.sopterm.makeawish.domain.user.User;
 import com.sopterm.makeawish.domain.wish.Wish;
@@ -33,12 +28,5 @@ public record WishRequestDTO(
 			.phoneNumber(phone)
 			.wisher(wisher)
 			.build();
-	}
-
-	public static LocalDateTime convertToTime(String date) {
-		val instant = Instant
-			.from(DateTimeFormatter.ISO_DATE_TIME.parse(date))
-			.atZone(ZoneId.of("Asia/Seoul"));
-		return LocalDateTime.from(instant);
 	}
 }
